requirements.txt
numpy
scipy
pycryptodome
soundfile